Set source rank for multi-node NCCL broadcast collectives.
Before this change, `root_rank` in `NcclManager` was inferred from the call to
`AddBroadcastSend`.  However, if the collective op is multi-node, then there
may not be a broadcast send call at a node.

This change enables passing the rank of the broadcast source in the context,
and adds some runtime checks to ensure consistency of the source rank.  If the
source rank in collective context is negative, then we assume this is a single
node collective and take the rank of the caller of broadcast send as the source
rank.

PiperOrigin-RevId: 262164533